requirements.txt
certifi==2019.3.9
Django==2.1.7
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
psycopg2==2.7.7
pytz==2018.9
simplejson==3.16.0
